Quito la opcion de eliminar solo dejo modificar por separado  y optimizo el codigo
